List permissions fix No. 3 in a series
So maybe there are two ways to create a list in shavar and maybe we use
one not the other. Let's be sure and set pbject acl's either way.

I blame someone else for this not being fixed the first time.

Allow Cloudfront user read access to objects when publishing them

This was a very obscure bug. When creating the lists in prod, the owner
of the list differs from that of the owner of the bucket. For this
reason, cloudfront can't serve the file without making it public read,
which we don't want. The solution is to explicitly grant the canonical
user id for the cloudfront identity read access to the object. This way,
the bucket policy remains in effect but we can still serve the content
via CloudFront.