Add unit tests for canonicalize
Add pytest as a dependency and pin it to the last version that supports
Python 2 (i.e., v4.6.9). Implement a unit test that validates the
canonicalization implementation, using some of the suggested test cases
from the Safe Browsing v2 API documentation with some adjustments and
additions. Also, implement a unit test that confirms that the return
type of canonicalize is str. Finally, add instructions for running the
unit tests in the README.