Remove whitespace surrounding urls in canonicalize
Use strip() to remove all leading and trailing whitespace from a url in
canonicalize(). Previously, a url with leading whitespace characters
would cause canonicalize() to fail with an AttributeError because the
regular expression used to extract the hostname and path could not match
it. Also, add two relevant test cases.