Fix bug in replacing consecutive slashes with one
Replace consecutive slashes found in the path of a url passed to
canonicalize with a single slash regardless of their position in the
path. Previously, this was done only for consecutive slashes in the
beginning of the path. Make sure this replacement does not happen in
query parameters. Modify one of the canonicalization test cases to
validate this behavior. Also, replace calls to re.subn with calls to
re.sub as we do not use the number of substitutions.