Percent-escape special characters in the whole url
Percent-escape characters that are ASCII <= 32, >= 127, # or % in both
the hostname and the path before returning a canonicalized url. This
behavior complies with the canonicalization rules described in the Safe
Browsing v2 documentation, as opposed to the previous implementation,
which only percent-escaped special characters in the path. Add some of
the suggested test cases from the Safe Browsing v2 documentation that
were previously failing because of this discrepancy.