Handle public and private PSL domains differently
Raise a ValueError in add_domain_to_list when the domain is in the
public (ICANN) section of the Public Suffix List, instead of silently
skipping the domain's addition to the list and logging. If the domain is
in the private section of the Public Suffix List, add it to the
blocklist. Bump the version of publicsuffixlist to 0.7.3, so that we can
use its "only_icann" kwarg. Add tests for both cases. Closes #102