Do not add domains twice in add_domain_to_list
Modify add_domain_to_list so that it returns False and does not add the
domain to the list if it is found in the previous_domains set.
test_add_domain_to_list_duplicate, which was previously failing, now
succeeds. Also, change the docstring of write_safebrowsing_blocklist to
reflect the fact that it receives a set of domains as an argument and
not a list.