Configure coverage reporting
Add pytest-cov as a dependency. Configure CircleCI to produce an XML
coverage report when running the tests and upload it to Codecov using
Codecov's bash uploader. Add a configuration file for Codecov to enable
status checks for commits and PRs without enforcing a coverage threshold
for now. Also, disable Codecov comments on PRs. Additionally, modify the
instruction for running tests in the README so that it also produces a
coverage report. Finally, add .coverage to gitignore.