Refactor new_data_to_publish_to_s3
Move the code that uses the configuration file to determine if a list
should be uploaded to S3 outside of `new_data_to_publish_to_s3` and into
`publish_to_cloud`. Rename `check_upload_remote_settings_config` to
`check_upload_config` and adjust it to work both for S3 and remote
settings. In `publish_to_cloud` make sure that
`new_data_to_publish_to_s3` is only called when `upload_to_s3` is True,
to avoid the cost of unnecessarily accessing S3.

In addition, when the key corresponding to a list does not yet exist on
S3, return True right away instead of creating a key with dummy list
contents and comparing against its checksum.

Also, fix a minor bug in getting the name of the S3 key from the
configuration file. Previously, if the section did not have the `s3_key`
option, `new_data_to_publish_to_s3` would raise a `NoOptionError`
instead of falling back to using the `output` option as the S3 key name.
Raise a ValueError when the `s3_key` option exists but is empty.

Finally, add a docstring to describe the functionality of
`new_data_to_publish_to_s3`.