Convert to bytes/string where appropriate
Make sure to encode string objects to bytes and decode byte objects to
strings wherever necessary, as a preparatory step for transitioning to
Python 3. More specifically:

  * Encode URLs before hashing them, as hashlib.sha256() expects a bytes
    object in Python 3
  * Convert chunk headers to bytes before writing them to a binary file
  * Decode data right after receiving it from urlopen
  * Decode the content of HTTP responses before printing it
  * Decode data right after reading it from a binary file
  * Encode the data used to mock urlopen in tests

Also, fix a minor bug by using the 'b' prefix at the start of each line
of a multi-line bytes literal instead of only using it for the first
line. Finally, omit the "utf-8" argument when encoding, as UTF-8 is the
default encoding.