Order domains in tracker blocklists
Sort the canonicalized domains alphanumerically before writing their
hashes to the safebrowsing list file. This ensures that any changes in
the order of domains in the input JSON lists will not cause unnecessary
list updates. Additionally, this is a prerequisite for porting the list
creation script to Python 3. Hash randomization is enabled by default
starting from Python 3.3 and this would result in a differently ordered
list being created every time the script is run.

Also, replace the previous_domains set with a previous_domain string.
Now that the domains are ordered, if there are duplicates they will be
consecutive. Finally, adjust the tests to the changes and make sure that
write_safebrowsing_blocklist tests check both the order of hashes and
the existence of duplicates.

Note: the dummytracker domains are exempt from ordering and remain at
the top of the list file.