Order domains in plugin blocklists
Sort the canonicalized domains alphanumerically before writing their
hashes to the safebrowsing list file. Replace the previous domains set
with a previous_domain string. Finally, adjust the tests to the changes
and make sure that process_plugin_blocklist tests check both the order
of hashes and the existence of duplicates.