Order domains in entity lists
Sort the canonicalized URLs of each entity alphanumerically before
writing their hashes to the safebrowsing list file. URLs are still
sorted based on the entity name and this commit introduces a secondary
ordering within each entity. Use a set to store canonicalized URLs
before sorting them, so that duplicates are eliminated. Finally, make
sure that process_entitylist tests check both the order of hashes and
the existence of duplicates.